+++ src/euroeval/prompt_templates/sentiment_classification.py
@@ -26,6 +26,7 @@ from ..languages import (
     PL,
     PT,
     SK,
+    SR,
     SV,
     UK,
 )
@@ -258,6 +259,16 @@ SENT_TEMPLATES: dict["Language", PromptConfig] = {
         default_prompt_template="Dokument: {text}\nSentiment: {label}",
         default_instruction_prompt="Dokument: {text}\n\nKlasifikujte pocit v "
         "dokumente. Odpovedzte so {labels_str}, a nič iné.",
+    ),
+    SR: PromptConfig(
+        default_prompt_label_mapping=dict(
+            positive="pozitivan", neutral="neutralan", negative="negativan"
+        ),
+        default_prompt_prefix="U nastavku su dokumenti i njihov sentiment, koji može "
+        "biti {labels_str}.",
+        default_prompt_template="Dokument: {text}\nSentiment: {label}",
+        default_instruction_prompt="Dokument: {text}\n\nKlasifikujte sentiment u "
+        "dokumentu. Odgovorite sa {labels_str}, i ništa drugo.",
     ),
     SV: PromptConfig(
         default_prompt_label_mapping=dict(