+++ src/scripts/create_mmlu.py
@@ -28,7 +28,21 @@ from datasets import Dataset, DatasetDict, Split, load_dataset
 from huggingface_hub import HfApi
 from sklearn.model_selection import train_test_split
 
-LANGUAGES = ["da", "de", "en", "es", "fr", "is", "it", "nl", "no", "pt", "sk", "sv"]
+LANGUAGES = [
+    "da",
+    "de",
+    "en",
+    "es",
+    "fr",
+    "is",
+    "it",
+    "nl",
+    "no",
+    "pt",
+    "sk",
+    "sr",
+    "sv",
+]
 
 
 def main() -> None: