+++ src/scripts/create_multi_wiki_qa.py
@@ -47,6 +47,7 @@ LANGUAGES = [
     "uk",
     "el",
     "bg",
+    "sr",
 ]