+++ src/scripts/create_scala.py
@@ -43,6 +43,7 @@ from load_ud_pos import (
     load_pldt_pos,
     load_ptdt_pos,
     load_skdt_pos,
+    load_srdt_pos,
     load_svdt_pos,
     load_ukdt_pos,
 )
@@ -77,6 +78,7 @@ def main() -> None:
         "uk": load_ukdt_pos,
         "el": load_eldt_pos,
         "bg": load_bgdt_pos,
+        "sr": load_srdt_pos,
     }
 
     # Set up the progress bar and iterate over the languages