+++ src/scripts/create_winogrande.py
@@ -47,6 +47,7 @@ LANGUAGES = [
     "pl",
     "pt",
     "sk",
+    "sr",
     "sv",
     "uk",
 ]