+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added support for Serbian 🇷🇸! This includes the sentiment classification dataset
+  MMS-sr, the linguistic acceptability dataset ScaLA-sr, the named entity recognition
+  dataset UNER-sr, the reading comprehension dataset MultiWikiQA-sr, the summarisation
+  dataset LR-Sum-sr, the knowledge dataset MMLU-sr, and the common-sense reasoning dataset
+  Winogrande-sr. This was contributed by @oliverkinch ✨
 - Added support for Bulgarian 🇧🇬! This includes the sentiment classification dataset
   Cinexio, the linguistic acceptability dataset ScaLA-bg, the named entity recognition
   dataset BG-NER-BSNLP, the reading comprehension dataset MultiWikiQA-bg, the knowledge